Trigger Nushell fallback to stdio on attempt to run local socket mode
If Nushell fails to launch a plugin with local socket mode, even if the
plugin advertises support for it, it automatically falls back to stdio
mode. This is so that if for some reason you're doing something strange
(like piping the data through something :wink:) or your operating system
hates you, it still might work.

We can trigger this fallback by just silently exiting when we encounter
the `--local-socket` arg. Nushell will treat this as something having
gone wrong, and launch with stdio.

It might be nice to support local socket mode anyway, since there are
some plugins that can't work without it, generally terminal interface
plugins.